+++ mod1/src/bezier_y.go
@@ -1,12 +1,12 @@
 package main
 
-func bezier_func_y(list bezier, tab *[width][height]float32) {
-	var t float32 = 0
-	var x float32
+func bezier_func_y(list bezier, tab *[width][height]float64) {
+	var t float64 = 0
+	var x float64
 
 	for list.start.x+x <= list.end.x {
-		t = float32(x) / float32((list.end.x - list.start.x))
-		tab[int(list.start.x+x)][int(list.start.y)] = float32(bez_deg3(list.start, list.end, list.po1, list.po2, t))
+		t = float64(x) / float64((list.end.x - list.start.x))
+		tab[int(list.start.x+x)][int(list.start.y)] = float64(bez_deg3(list.start, list.end, list.po1, list.po2, t))
 		x++
 	}
 }
